Tests for error cases
