+++ src/DI.Specification.Tests/Fakes/ClassWithAmbiguousCtors.cs
@@ -1,8 +1,6 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System.Collections.Generic;
-
 namespace Microsoft.Extensions.DependencyInjection.Specification.Fakes
 {
     public class ClassWithAmbiguousCtors
@@ -29,32 +27,6 @@ namespace Microsoft.Extensions.DependencyInjection.Specification.Fakes
             Data2 = data2;
 
             CtorUsed = "IFakeService, string, string";
-        }
-
-        public IFakeService FakeService { get; }
-
-        public string Data1 { get; }
-
-        public int Data2 { get; }
-        public string CtorUsed { get; set; }
-    }
-
-    public class ClassWithAmbiguousCtorsAndAttribute
-    {
-        public ClassWithAmbiguousCtorsAndAttribute(string data)
-        {
-            CtorUsed = "string";
-        }
-
-        [ActivatorUtilitiesConstructor]
-        public ClassWithAmbiguousCtorsAndAttribute(IFakeService service, string data)
-        {
-            CtorUsed = "IFakeService, string";
-        }
-
-        public ClassWithAmbiguousCtorsAndAttribute(IFakeService service,  IFakeService service2, string data)
-        {
-            CtorUsed = "IFakeService, IFakeService, string";
         }
 
         public IFakeService FakeService { get; }
